requirements.txt
asgiref==3.11.1
Django==6.0.3
sqlparse==0.5.5
tzdata==2025.3
python-decouple
Pillow
djangorestframework
markdown
django-filter

drf-yasg
djangorestframework-simplejwt
django-mail-templated


black
flake8